refactor(coverage): isolate the untestable-headless residue (#103)
Split the structurally-untestable glue out of two well-covered files so
codecov's denominator only counts headlessly-reachable code, per the
existing "keeps the % honest" convention (main.rs, tui/mod.rs):

- runtime.rs -> runtime/mod.rs (RunConfig, boot-capacity math, summarize
  - all unit-tested) + runtime/driver.rs (tokio runtime, run/run_async,
  reducer_task, headless_loop, compute_boot_capacities - block_on +
  ctrl_c + socket bind, unreachable by any headless test). driver.rs is
  codecov-ignored; `runtime::run` is re-exported so main.rs is untouched.
- tui/renderer.rs: Term + setup_terminal/teardown_terminal move into the
  already-ignored tui/mod.rs - raw mode + alternate screen belong with
  the event loop that owns them, and renderer.rs (draw_scene/flush) now
  carries no real-TTY code.

Accepted as residue per the issue: the claude_code.rs select! abort arms
(forcing one infinite loop to win is racy) and the 2 uncaptured hook-shim
lines (behaviorally covered by tests/shim.rs as a subprocess).

Multi-agent review fixes folded in:
- claude-security-review.yml paths filter: every entry except
  pixtuoid-hook/** was missing the src/ segment, so the workflow has
  never fired on its intended install/runtime/source changes
  (pre-existing since bf5d2af); fixed + runtime entry now targets the
  split (src/runtime/**).
- registry.rs + pixel_painter/mod.rs doc comments updated for the move.
- CLAUDE.md notes the semver gate covers pixtuoid-core only (the
  pixtuoid lib target is internal-facing, not a semver surface - this
  move deletes pub paths tui::renderer::{Term,setup_terminal,
  teardown_terminal}, technically major, zero consumers).

Behavior-preserving move - no logic changes; 1014/1014 + preflight green.
